requirements.txt
bs4==0.0.1
requests==2.23.0
pandas==1.0.3
mecab-python3==1.0.3
seaborn==0.11.1
matplotlib==3.2.1
